update all tests to use new style.
The new pull(a,b,c) style is more conveinient that a.pipe(b).pipe(c)
not only is there less clutter, but also, syntax errors are more helpful.
in the old way, if you missed a bracket, the error message doesn't point you
to the place where the bracket should be - instead it points to where the
first unexpected token was found - with pull(,,,), if you miss out a comma,
you'll most probably get a syntax error on the next line.